Fix invalid required collat amnt check
The check in Request.hs file containsRequiredCollateralAmount used the opposite value comparison.
It checked that the collateral in the input (Request script UTxO) is more or the same compared to the ongoing Collateral script UTxO. This means that the Lender used to be able to take the whole collateral when providing the loan.
Implemented change is to reverse comparison check and to compare the ongoing locked collateral to the collateral value noted in the Request datum instead of the currently locked value.